refactoring: wrap_model->Model, wrap_dataset->Dataset
