+++ python-client/docs/guides/tutorials/huggingface/BertForSequenceClassification.md
@@ -13,7 +13,7 @@ import pandas as pd
 from dateutil import parser
 from scipy import special
 from transformers import BertTokenizer, BertForSequenceClassification
-from giskard import wrap_model, wrap_dataset
+from giskard import Model, Dataset
 ```
 
 ## Wrap dataset
@@ -107,7 +107,7 @@ excluded_category = [idx_to_cat[i] for i in [11, 12, 13]]
 data_filtered = data_filtered[-data_filtered["Target"].isin(excluded_category)]
 ```
 ```python
-wrapped_dataset = wrap_dataset(data_filtered.head(5), 
+wrapped_dataset = Dataset(data_filtered.head(5), 
                                name="test dataset", 
                                target="Target",
                                cat_columns=['Week_day', 'Month'])
@@ -136,7 +136,7 @@ def preprocessing_func(test_dataset):
     return X_test_tokenized
 ```
 ```python
-wrapped_model = wrap_model(
+wrapped_model = Model(
     name=model_name,
     model=model,
     feature_names=['Content'],
