+++ python-client/docs/guides/tutorials/huggingface/pytorch_pipeline.md
@@ -5,7 +5,7 @@
 # For the complete tutorial, check: https://huggingface.co/docs/transformers/tasks/sequence_classification
 import pandas as pd
 from transformers import pipeline
-from giskard import wrap_model, wrap_dataset
+from giskard import Model, Dataset
 ```
 ## Wrap dataset
 ```python
@@ -18,7 +18,7 @@ raw_data = {
 test_df = pd.DataFrame(raw_data, columns=["text", "label"], index=[0])
 ```
 ```python
-wrapped_dataset = wrap_dataset(test_df, 
+wrapped_dataset = Dataset(test_df, 
                                name="test dataset", 
                                target="label")
 ```
@@ -37,7 +37,7 @@ def my_preproccessing_function(df):
 
 ```
 ```python
-wrapped_model = wrap_model(
+wrapped_model = Model(
     name="stevhliu/my_awesome_model",
     model=my_classifier,
     feature_names=feature_names,
