+++ python-client/docs/reference/datasets/index.rst
@@ -1,9 +1,7 @@
-Datasets
+Dataset
 ============
 Currently we only support **tabular** or **NLP** data types. Your data must be based on a :code:`pandas.DataFrame` and wrapped
-with our own :class:`~.giskard.Dataset` class. The wrapping is done via the :func:`~.giskard.wrap_dataset` factory method.
-
-.. autofunction:: giskard.wrap_dataset
+with our own :class:`~.giskard.Dataset` class.
 
 .. autoclass:: giskard.Dataset
