+++ python-client/giskard/__init__.py
@@ -5,7 +5,6 @@ import sys
 
 from giskard.client.giskard_client import GiskardClient
 from giskard.client.project import Project
-from giskard.datasets import wrap_dataset
 from giskard.datasets.base import Dataset
 from giskard.ml_worker.core.suite import Suite, SuiteInput
 from giskard.ml_worker.core.test_result import TestResult
@@ -28,9 +27,6 @@ from giskard.ml_worker.testing.tests.performance import AucTest, test_mae, test_
     test_diff_reference_actual_f1
 from giskard.ml_worker.testing.tests.statistic import test_right_label, test_output_in_range, test_disparate_impact
 from giskard.ml_worker.utils.logging import configure_logging
-from giskard.models import wrap_model, model_from_catboost, model_from_huggingface, model_from_tensorflow, \
-    model_from_pytorch, model_from_sklearn
-from giskard.models.base import WrapperModel, CustomModel, MLFlowBasedModel, BaseModel, CloudpickleBasedModel
 from giskard.models.automodel import Model
 
 from .scanner import scan
@@ -54,21 +50,9 @@ __version__: str = get_version()
 __all__ = [
     'SingleTestResult',
     'Project',
-    'wrap_dataset',
     'Dataset',
     'GiskardClient',
     'test',
-    'wrap_model',
-    'model_from_catboost',
-    'model_from_sklearn',
-    'model_from_pytorch',
-    'model_from_tensorflow',
-    'model_from_huggingface',
-    'BaseModel',
-    'WrapperModel',
-    'MLFlowBasedModel',
-    'CloudpickleBasedModel',
-    'CustomModel',
     'Model',
     'Suite',
     'test_drift_psi',
