+++ python-client/tests/test_dataset.py
@@ -1,4 +1,4 @@
-from giskard.datasets import wrap_dataset
+from giskard.datasets import Dataset
 from giskard.datasets.base import Dataset
 import pandas as pd
 import pytest
@@ -14,7 +14,7 @@ nonvalid_df = pd.DataFrame({"categorical_column": [['turtle'], ['crocodile'], ['
 
 
 def test_factory():
-    my_dataset = wrap_dataset(valid_df)
+    my_dataset = Dataset(valid_df)
     assert isinstance(my_dataset, Dataset)
 
 
@@ -24,20 +24,20 @@ def test_valid_df_column_types():
             UserWarning,
             match="You did not provide the optional argument 'target'\. 'target' is the column name " # noqa
                   "in df corresponding to the actual target variable \(ground truth\)\."): # noqa
-        my_dataset = wrap_dataset(valid_df)
+        my_dataset = Dataset(valid_df)
     assert my_dataset.column_types == {'categorical_column': 'category', 'text_column': 'text', 'numeric_column': 'numeric'}
 
     # Option 1: column_types is provided
-    my_dataset = wrap_dataset(valid_df, column_types=valid_df_column_types)
+    my_dataset = Dataset(valid_df, column_types=valid_df_column_types)
     assert my_dataset.column_types == valid_df_column_types
 
     # Option 2: cat_columns is provided
     cat_columns = ['categorical_column']
-    my_dataset = wrap_dataset(valid_df, cat_columns=cat_columns)
+    my_dataset = Dataset(valid_df, cat_columns=cat_columns)
     assert my_dataset.column_types == valid_df_column_types
 
     # Option 3: infer_column_types is provided
-    my_dataset = wrap_dataset(valid_df)
+    my_dataset = Dataset(valid_df)
     assert my_dataset.column_types == valid_df_column_types
 
 
@@ -47,14 +47,14 @@ def test_nonvalid_df_column_types():
             TypeError,
             match="The following columns in your df: \['categorical_column', 'text_column'\] are not hashable\. " # noqa
                   "We currently support only hashable column types such as int, bool, str, tuple and not list or dict\."): # noqa
-        wrap_dataset(nonvalid_df)
+        Dataset(nonvalid_df)
 
     # Option 1: column_types is provided
     with pytest.raises(
             TypeError,
             match="The following columns in your df: \['categorical_column', 'text_column'\] are not hashable\. " # noqa
                   "We currently support only hashable column types such as int, bool, str, tuple and not list or dict\."): # noqa
-        wrap_dataset(nonvalid_df, column_types=valid_df_column_types)
+        Dataset(nonvalid_df, column_types=valid_df_column_types)
 
     # Option 2: cat_columns is provided
     cat_columns = ['categorical_column']
@@ -62,11 +62,11 @@ def test_nonvalid_df_column_types():
             TypeError,
             match="The following columns in your df: \['categorical_column', 'text_column'\] are not hashable\. " # noqa
                   "We currently support only hashable column types such as int, bool, str, tuple and not list or dict\."): # noqa
-        wrap_dataset(nonvalid_df, cat_columns=cat_columns)
+        Dataset(nonvalid_df, cat_columns=cat_columns)
 
     # Option 3: infer_column_types is provided
     with pytest.raises(
             TypeError,
             match="The following columns in your df: \['categorical_column', 'text_column'\] are not hashable\. " # noqa
                   "We currently support only hashable column types such as int, bool, str, tuple and not list or dict\."): # noqa
-        wrap_dataset(nonvalid_df)
+        Dataset(nonvalid_df)
